requirements.txt
streamlit==1.22.0
   pandas==1.5.3
   Pillow==9.5.0
   PyPDF2==3.0.1